Split vulkan_handle_wrappers.h
Split vulkan_handle_wrappers.h into two separate files:
- vulkan_state_info.h to define types used to store state information.
- vulkan_handle_wrappers.h that only defines handle wrappers; handle
  wrappers reference types from vulkan_state_info.h
