+++ android/framework/encode/CMakeLists.txt
@@ -17,6 +17,7 @@ target_sources(gfxrecon_encode
                    ${GFXRECON_SOURCE_DIR}/framework/encode/trace_manager.h
                    ${GFXRECON_SOURCE_DIR}/framework/encode/trace_manager.cpp
                    ${GFXRECON_SOURCE_DIR}/framework/encode/vulkan_handle_wrappers.h
+                   ${GFXRECON_SOURCE_DIR}/framework/encode/vulkan_state_info.h
                    ${GFXRECON_SOURCE_DIR}/framework/encode/vulkan_state_table.h
                    ${GFXRECON_SOURCE_DIR}/framework/encode/vulkan_state_tracker.h
                    ${GFXRECON_SOURCE_DIR}/framework/encode/vulkan_state_tracker.cpp
