+++ framework/encode/vulkan_handle_wrappers.h
@@ -14,9 +14,10 @@
 ** limitations under the License.
 */
 
-#ifndef GFXRECON_ENCODE_HANDLE_WRAPPERS_H
-#define GFXRECON_ENCODE_HANDLE_WRAPPERS_H
+#ifndef GFXRECON_ENCODE_VULKAN_HANDLE_WRAPPERS_H
+#define GFXRECON_ENCODE_VULKAN_HANDLE_WRAPPERS_H
 
+#include "encode/vulkan_state_info.h"
 #include "format/format.h"
 #include "util/defines.h"
 #include "util/memory_output_stream.h"
@@ -30,75 +31,6 @@
 
 GFXRECON_BEGIN_NAMESPACE(gfxrecon)
 GFXRECON_BEGIN_NAMESPACE(encode)
-
-//
-// Types for state tracking.
-//
-
-typedef std::shared_ptr<util::MemoryOutputStream> CreateParameters;
-
-// Active query state information stored with VkQueryPool handle.
-struct QueryInfo
-{
-    static const uint32_t kInvalidIndex = std::numeric_limits<uint32_t>::max();
-
-    VkQueryControlFlags flags{ 0 };
-    uint32_t            index{ kInvalidIndex };           // Pool index for active query.
-    VkCommandBuffer     command_buffer{ VK_NULL_HANDLE }; // Command buffer for query begin.
-    format::HandleId    command_buffer_id{ 0 };
-    VkRenderPass        render_pass{ VK_NULL_HANDLE }; // Optional render pass containing query.
-    format::HandleId    render_pass_id{ 0 };
-};
-
-struct DescriptorBindingInfo
-{
-    uint32_t         binding_index{ 0 };
-    uint32_t         count{ 0 };
-    VkDescriptorType type;
-};
-
-struct DescriptorInfo
-{
-    VkDescriptorType                          type;
-    uint32_t                                  count{ 0 };
-    std::unique_ptr<bool[]>                   written;
-    std::unique_ptr<VkDescriptorImageInfo[]>  images;
-    std::unique_ptr<VkDescriptorBufferInfo[]> buffers;
-    std::unique_ptr<VkBufferView[]>           texel_buffer_views;
-};
-
-// VkDescriptorSetLayout create info stored with VkPipelineLayout handle.
-struct DescriptorSetLayoutInfo
-{
-    VkDescriptorSetLayout handle{ VK_NULL_HANDLE };
-    format::HandleId      handle_id{ 0 };
-    format::ApiCallId     create_call_id{ format::ApiCallId::ApiCall_Unknown };
-    CreateParameters      create_parameters;
-};
-
-// Create info for all descriptor set layouts required by a pipeline layout.
-// Referenced with a shared pointer by VkPipelineLayout and VkPipeline handles.
-struct PipelineLayoutDependencies
-{
-    std::vector<DescriptorSetLayoutInfo> layouts;
-};
-
-// VkShaderModule create info stored with VkPipeline handle.
-struct ShaderModuleInfo
-{
-    VkShaderModule    handle{ VK_NULL_HANDLE };
-    format::HandleId  handle_id{ 0 };
-    format::ApiCallId create_call_id{ format::ApiCallId::ApiCall_Unknown };
-    CreateParameters  create_parameters;
-};
-
-struct ImageAcquiredInfo
-{
-    bool        is_acquired{ true };
-    VkSemaphore acquired_semaphore{ VK_NULL_HANDLE };
-    VkFence     acquired_fence{ VK_NULL_HANDLE };
-    VkQueue     last_presented_queue{ VK_NULL_HANDLE };
-};
 
 //
 // Handle wrappers for storing object state information with object handles.
@@ -399,4 +331,4 @@ struct AccelerationStructureNVWrapper : public HandleWrapper<VkAccelerationStruc
 GFXRECON_END_NAMESPACE(encode)
 GFXRECON_END_NAMESPACE(gfxrecon)
 
-#endif
+#endif // GFXRECON_ENCODE_VULKAN_HANDLE_WRAPPERS_H
